uc1-2 fortsat + lidt til præsentation

+++ src/domain/Profil.java
@@ -2,12 +2,12 @@ package domain;
 
 public interface Profil {
 
-	public double getKundeNummer();
-	public void setKundeNummer(double kundeNummer);
+	public long getKundeNummer();
+	public void setKundeNummer(long kundeNummer);
 	public String getFuldtNavn();
 	public void setFuldtNavn(String fuldtNavn);
 	public String getEmail();
 	public void setEmail(String email);
-	public double getTlfNummer();
-	public void setTlfNummer(double tlfNummer);
+	public long getTlfNummer();
+	public void setTlfNummer(long tlfNummer);
 }
